leet code solution question8
